requirements.txt
attrs==23.2.0
graphviz==0.20.3
hydra-core==1.3.2
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.2
pandaserd==0.1.5
pymongo==4.7.2
python-slugify==8.0.4
Requests==2.32.2
seaborn==0.13.2
sentence_transformers==2.7.0
setuptools==59.6.0
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.4
tabulate==0.9.0
cattrs==23.2.3